Implements RequireJS as an AMD manager for our Backbone.js admin application.
The entry point to the application is `assets/js/config.js` which configures jQuery, Underscore, Backbone and Modernizr as dependent libraries and `assets/js/main.js` as the application loader.

The `main.js` file initiates the Router and Menu view, which are loaded and initialised ready to start the app.